# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Supply Chain & Manufacturing Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalSkus: number;
  totalInventoryValue: number;
  reorderAlertCount: number;
  activeSuppliers: number;
  openOrders: number;
  pendingOrderValue: number;
  oeePercent: number;
  forecastAccuracy: number;
  fillRate: number;
  otifPercent: number;
}