# lib/api-zod/src/generated/types/demandRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Supply Chain & Manufacturing Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DemandRecord {
  id: number;
  productName: string;
  period: string;
  actualDemand: number;
  forecastedDemand: number;
  createdAt: string;
}