# lib/api-zod/src/generated/types/forecastResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Supply Chain & Manufacturing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ForecastResultTrend } from './forecastResultTrend';

export interface ForecastResult {
  productName: string;
  mape: number;
  mad: number;
  forecastAccuracy: number;
  nextPeriodForecast: number;
  trend: ForecastResultTrend;
}