# lib/api-zod/src/generated/types/inventoryHealth.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Supply Chain & Manufacturing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryBreakdown } from './categoryBreakdown';

export interface InventoryHealth {
  avgTurnoverRate: number;
  avgDaysOfSupply: number;
  overstockCount: number;
  stockoutCount: number;
  healthyCount: number;
  categoryBreakdown: CategoryBreakdown[];
}